requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.83
botocore==1.31.83
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.4.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
numpy==1.26.2
packaging==23.2
phonenumbers==8.13.24
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
s3transfer==0.7.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
celery==5.2.7
flower==1.2.0
redis==4.4.0